Refactor GAN
We trade-off training speed for code complexity. Before we compiled
the training functions ourself. This allowed training generator and
discriminator in one single training pass. Now, we rely on the keras
training functions. This is definitely the sane way from a software
engineer perspective. It would be cool, if keras had the
possibility to have different optimizers for different losses.

Thanks, to this commit a lot of nasty code is removed (mostly internal
keras code, that was copied and slightly adapted).